眼底疾病分类/requirements.txt
certifi==2018.1.18
chardet==3.0.4
cycler==0.11.0
fonttools==4.33.3
httplib2==0.9.2
idna==2.6
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.21.6
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
PyGObject==3.26.1
pyparsing==3.0.9
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==1.0.2
scipy==1.7.3
six==1.11.0
systemd-python==234
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.22
